feat: split index into components
