+++ frontend/assets/css/main.css
@@ -1,1 +1,4 @@
 @import "tailwindcss";
+.content-card {
+    @apply bg-white rounded-2xl shadow-xl p-4 md:p-6;
+}
